requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-mpesa==2.0.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
Markdown==3.5.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.1